requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
beanie==1.23.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
google-auth-oauthlib==1.1.0
google-auth==2.23.4
websockets==12.0
